backend/requirements.txt
# Core Flask / FastAPI
Flask>=2.3.3
flask-cors>=4.0.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0

# Scientific Computing
numpy>=1.26.4
scipy>=1.11.1
pandas>=2.0.3
rasterio>=1.3.8
geopandas>=0.13.2
shapely>=2.0.1
pyproj>=3.6.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=1.7.6
lightgbm>=4.0.0
joblib>=1.3.1

# Async HTTP
aiohttp>=3.8.5
aiofiles>=23.1.0

# Earth Observation (SAR Engine — Phase 4)
earthengine-api>=1.1.0

# Performance
numba>=0.57.1
ujson>=5.8.0

# Monitoring & Logging
prometheus-flask-exporter>=0.22.4
structlog>=23.1.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
requests>=2.31.0
python-multipart>=0.0.6
psycopg2-binary>=2.9.9

# Production
gunicorn>=21.2.0
eventlet>=0.33.3

# Development
black>=23.7.0
flake8>=6.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
